fix user side

+++ server/controllers/goal.controller.js
@@ -1,20 +1,26 @@
 const Goal = require('../models/goal.model');
 
 module.exports.getAllGoals = (req, res) => {
-    Goal.find().sort({ createdAt: -1 })
+    Goal.find({ user_id: req.userId }) 
+        .sort({ createdAt: -1 })
         .then(goals => res.json(goals))
         .catch(err => res.status(400).json(err));
 };
 
 module.exports.createGoal = (req, res) => {
-    Goal.create(req.body)
+    const newGoal = {
+        ...req.body,
+        user_id: req.userId
+    };
+
+    Goal.create(newGoal)
         .then(goal => res.json(goal))
         .catch(err => res.status(400).json(err));
 };
 
 module.exports.updateGoal = (req, res) => {
     Goal.findOneAndUpdate(
-        { _id: req.params.id }, 
+        { _id: req.params.id, user_id: req.userId }, 
         req.body, 
         { new: true, runValidators: true }
     )
@@ -23,7 +29,7 @@ module.exports.updateGoal = (req, res) => {
 };
 
 module.exports.deleteGoal = (req, res) => {
-    Goal.deleteOne({ _id: req.params.id })
+    Goal.deleteOne({ _id: req.params.id, user_id: req.userId })
         .then(result => res.json(result))
         .catch(err => res.status(400).json(err));
 };
